requirements.txt
python-dotenv==1.0.0
supabase==2.10.0
httpx==0.27.0
pandas==2.1.0